requirements.txt
Django==4.1.7
djangorestframework==3.12
psycopg2-binary==2.9.1
drf-yasg==1.20.0
pytz==2022.7.1